homework_11/pr5/worker/requirements.txt
kafka-python
torch
transformers